Fixed animation for big image
Big image now moves off the right side of the screen when scrolling up to read the paragraph. Added extra lines to set right property of the big image when performing the slide in and slide out animations.

Added event listener for scrolling on the page. Big image slides out when scrolled up a certain point in order to read the paragraph.