Update display_character.js
Resizing of big image depends on the proportions of the entry image. If the height is bigger than the width, then the resize height with be set and the resize width with depend on the resize height. Similarly, if the width is bigger than the height, the resize width is set and the resize height is dependent on the resize width.